Create Bluetooth_communication&security_feature
In this file I established a connection between my phone and bluetooth hc-05 module. I also generate an array of different numbers between 0 and 11. I will use this random array as a security method. Every index from my array correspons to a specific place in the matrix displayed in my android application; because everytime we will have numbers arranged different in the array, the matrix will also change and in this way, another person will not be able to steal yout pin code.